utils/requirements.txt
python-dotenv==0.19.0
PyYAML==5.4.1
snowflake-connector-python==2.5.1